[DT-709] Build initial payments UI

+++ payments/base/payment-manager/src/main/java/cm/aptoide/pt/payment_manager/manager/domain/PurchaseRequest.kt
@@ -9,7 +9,7 @@ data class PurchaseRequest(
   val callbackUrl: String?,
   val orderReference: String?,
   val signature: String?,
-  val value: Int?,
+  val value: Double?,
   val currency: String?,
   val oemId: String?,
 )
